Project Refactor Re-organized project for multiple modules
