+++ BeigeBoxPlugin/src/DataProcessor.cpp
@@ -1,7 +1,7 @@
 #define MAX_BUFFER 20480
 
-#include "DataProcessor.h"
-#include "DataUtil.h"
+#include "include/DataProcessor.h"
+#include "include/DataUtil.h"
 
 #include <fstream>
