+++ BeigeBoxPlugin/src/Logger.cpp
@@ -2,7 +2,7 @@
 // Created by Walter on 6/23/2024.
 //
 
-#include "Logger.h"
+#include "include/Logger.h"
 #include <iostream>
 #include <iomanip>
 #include <fstream>
