+++ BeigeBoxPlugin/src/include/Logger.h
