+++ BeigeBoxServer/CMakeLists.txt
@@ -1,12 +1,3 @@
-cmake_minimum_required(VERSION ${CMAKE_MAJOR_VERSION}.${CMAKE_MINOR_VERSION})
-set(PROJECT_NAME "BBServer")
-set(PROJECT_SOURCE_DIR ..)
-set(CMAKE_LIBRARY_TYPE SHARED)
-set(CMAKE_CXX_STANDARD ${CMAKE_LANGUAGE_VERSION})
-set(BOOST_USE_STATIC_LIBS ON)
-set(Boost_USE_MULTITHREADED FALSE)
-set(Boost_DEBUG ON)
-
 set(XPLANE_INCLUDE_DIR "SDK/xplane")
 
 if(WIN32)
@@ -27,49 +18,36 @@ if(NOT Boost_FOUND)
     message("Could not find boost at: ${BOOST_INCLUDE_DIR}")
 endif ()
 
-project(${PROJECT_NAME})
+project(BBServer)
 link_directories(../SDK/Libraries)
 
 include_directories(${XPLANE_INCLUDE_DIR})
-include_directories(src)
-include(src/CMakeLists.txt)
 
 
 if (MSVC)
     set(CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS ON)
 endif()
 
+add_executable(${PROJECT_NAME} src/main.cpp)
 
 if(WIN32)
     message("Building for windows")
-    add_definitions(-DIBM=1 -DBOOST_ALL_DYN_LINK=1)
-    option(BUILD_SHARED_LIBS "Build using shared libraries" ON)
-    target_link_libraries( bbserver
-            -static -static-libgcc -static-libstdc++ "../SDK/Libraries/Win/XPLM_64.lib"
-            -static -static-libgcc -static-libstdc++ "../SDK/Libraries/Win/XPWidgets_64.lib"
-            -static -static-libgcc -static-libstdc++ "../SDK/cwic/lib/cwic_lib.dll"
 
-            "ws2_32")
 elseif (APPLE)
-    add_definitions(-DAPL=1)
+    message("Build for apple")
 else ()
-    add_definitions(-DLIN=1)
-    set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY "..")
-    set(CMAKE_LIBRARY_OUTPUT_DIRECTORY "..")
-    set(CMAKE_RUNTIME_OUTPUT_DIRECTORY "..")
     message("Creating build for Linux")
     message(..)
-    target_link_libraries( bbserver PRIVATE "${PROJECT_SOURCE_DIR}/SDK/Libraries/libcwic_lib.so"
-            -shared -rdynamic -nodefaultlibs -undefined_warning "${PROJECT_SOURCE_DIR}/SDK/Libraries/libXPLM_64.so")
+
 endif (WIN32)
 
-set_target_properties(bbserver PROPERTIES
+set_target_properties(${PROJECT_NAME} PROPERTIES
         PREFIX ""
         OUTPUT_NAME "../server")
 
 
 if(WIN32)
-    install(TARGETS bbserver RUNTIME DESTINATION ../BeigeBox/win_x64)
+    install(TARGETS ${PROJECT_NAME} DESTINATION ../BeigeBox/win_x64)
 else()
-    install(TARGETS bbserver LIBRARY DESTINATION "${PROJECT_SOURCE_DIR}/BeigeBox/lin_x64/" )
+    install(TARGETS ${PROJECT_NAME} LIBRARY DESTINATION "${PROJECT_SOURCE_DIR}/BeigeBox/lin_x64/" )
 endif (WIN32)
